requirements.txt
django
djangorestframework
Django 5.2.7
djangorestframework 3.16.1